Initial commit for output plugin for mdm-statsd
The current plugin takes a record from fluentd with the required
information to send an MDM metric. The record is processed and a json
string is build to add the metrics namespace, dimensions and metric name
to be used by Geneva. Then it's sent via UDP to the statsd front end for
MDM.